# iqs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IQS Accreditation & Training API",
    "version": "1.0.0",
    "description": "API for school accreditation, evaluator management, training, messaging, and file handling.",
    "contact": {
      "email": "support@iqs-accreditation.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    { 
      "url": "http://localhost:5000",
      "description": "Development server" 
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [ { "bearerAuth": [] } ],
  "paths": {
...
    "/api/message/send": {
      "post": {
        "tags": ["Messaging"],
        "summary": "Send message",
        "description": "Send message to another user",
        "operationId": "sendMessage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "receiver_id": { 
                    "type": "integer",
                    "example": 789 
                  },
                  "message": { 
                    "type": "string",
                    "example": "Hello, when is our next meeting?" 
                  }
                },
                "required": ["receiver_id", "message"]
              }
            }
          }
        },
        "responses": { 
          "200": { 
            "description": "Message sent" 
          },
          "400": {
            "description": "Invalid receiver ID"
          }
        }
      }
    },
...
  }
}